backend_pathway/requirements_mac.txt
pathway
google-generativeai
fastapi
uvicorn
python-dotenv
